Adds Hamilton and feature engineering section
Hamilton was created to help wrangle a feature engineering code base.
It forces decoupling of feature transform logic from materialization, and results
in code that is always unit testable, reusable, and documentation friendly.  

I didn't see an appropriate section to add it, so I created a feature engineering
section - putting it in feature stores wouldn't be the right place.